feat(fame-server): update dependencies

+++ fame-server/src/main/java/com/zbw/fame/controller/admin/ArticleController.java
@@ -84,7 +84,7 @@ public class ArticleController {
      * @return {@see Integer}
      */
     @GetMapping("count")
-    public RestResponse<Integer> count() {
+    public RestResponse<Long> count() {
         return RestResponse.ok(articleService.count());
     }
 }
