+++ fame-server/src/main/java/com/zbw/fame/model/dto/ArticleDetailDto.java
@@ -37,7 +37,7 @@ public class ArticleDetailDto {
 
     private Boolean allowComment;
 
-    private Integer commentCount;
+    private Long commentCount;
 
     private Category category;
